+++ src/core/errors.ts
@@ -1,0 +1,11 @@
+export class DevFlowError extends Error {
+  constructor(
+    message: string,
+    public code: string,
+    public suggestions: string[] = []
+  ) {
+    super(message);
+    this.name = 'DevFlowError';
+    Object.setPrototypeOf(this, DevFlowError.prototype);
+  }
+}